+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+- No changes yet.
+
 ## [1.23.0] - 2020-03-31
 ### Added
 - Support opting out of the `omitempty` JSON option by adding `!omitempty` to the
@@ -330,6 +333,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Initial release.
 
+[Unreleased]: https://github.com/thriftrw/thriftrw-go/compare/v1.23.0...HEAD
 [1.23.0]: https://github.com/thriftrw/thriftrw-go/compare/v1.22.0...v1.23.0
 [1.22.0]: https://github.com/thriftrw/thriftrw-go/compare/v1.21.0...v1.22.0
 [1.21.0]: https://github.com/thriftrw/thriftrw-go/compare/v1.20.2...v1.21.0